requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
certifi==2024.8.30
distro==1.9.0
frozenlist==1.5.0
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
typing_extensions==4.12.2
yarl==1.18.3
